Remove launchHttpServer from Application class and expose it in index
